extract common code to the base class
